Moves `FormattedCellData` etc to its own header file.
